requirements.txt
firebase-admin
edge-tts
yt-dlp
pydub
google-generativeai